add icon to readme, add readme.txt to gitignore

+++ src/mappers/mapper72.rs
@@ -298,8 +298,8 @@ impl Mapper for Mapper72 {
 
     fn save_mapper_registers(&self, _cart: &Cartridge) -> Vec<u8> {
         let mut state = vec![self.preg, self.creg, self.variant as u8, self.prev, self.playing as u8];
-        state.extend_from_slice(&self.current_track.to_le_bytes());
-        state.extend_from_slice(&self.current_idx.to_le_bytes());
+        state.extend_from_slice(&(self.current_track as u64).to_le_bytes());
+        state.extend_from_slice(&(self.current_idx as u64).to_le_bytes());
         state
     }
 
@@ -317,11 +317,11 @@ impl Mapper for Mapper72 {
             start += 5;
             let mut track_bytes = [0u8; 8];
             track_bytes.copy_from_slice(&state[start..start + 8]);
-            self.current_track = usize::from_le_bytes(track_bytes);
+            self.current_track = u64::from_le_bytes(track_bytes) as usize;
             start += 8;
             let mut idx_bytes = [0u8; 8];
             idx_bytes.copy_from_slice(&state[start..start + 8]);
-            self.current_idx = usize::from_le_bytes(idx_bytes);
+            self.current_idx = u64::from_le_bytes(idx_bytes) as usize;
             start += 8;
         }
         start
